Optional catch-all segment 사용. /game, /game/121, /game/adf, /game/121/234 모두 붙잡아 처리할 수 있도록
CreateGame 버튼 따로 뺌. fetch로 받아와서 route할수 있게
