+++ front/app/main/@main/social/page.tsx
@@ -2,10 +2,8 @@ import ChatLayout from "@/components/Chat/ChatLayout";
 
 export default function Main() {
     return (
-        <>
-            <div className="flex h-full flex-col text-4xl font-extrabold">
-                <ChatLayout />
-            </div>
-        </>
+        <div className="flex h-full flex-col text-4xl font-extrabold">
+            <ChatLayout />
+        </div>
     );
 }
